Version 0.4 - UI Update
1) Optimized UI for 1080p:

Changed UI sprites scale and optimized them for 1920x1080 resolution.

2) Added screen safe area:

Now UI is placed in the safe area of a device screen.

3) Changed leaderboard design:

Now player's line has a different color in leaderboard.

4) Deleted exit button:

Now there is only the start button on the main menu screen.

5) Changed camera size:

Optimized camera size to make 2D sprites look better with 1080p and make the game more playable in the portrait mode.

Other changes:

6) Changed the size of the prefabs and the gameobjects on the scene.

7) Deleted a one prefab of obstacles.

8) Fixed a bug causing the save problem.

+++ Assets/Scripts/GameManager.cs
@@ -39,8 +39,6 @@ public class GameManager : MonoBehaviour
         }
 
         bestScore = PlayerPrefs.GetInt("bestScore");
-
-
     }
 
     public void ScoreUpdate()
@@ -89,8 +87,35 @@ public class GameManager : MonoBehaviour
         }
     }
 
-    private void OnApplicationQuit()
+    private void OnApplicationFocus(bool focus)
     {
-        PlayerPrefs.DeleteKey("isCameraMoved");
+        isGameOnPause = !focus;
+
+        if (isGameOnPause)
+        {
+            PlayerPrefs.DeleteKey("isCameraMoved");
+            Debug.Log("Game is unfocused");
+        }
+        else
+        {
+            PlayerPrefs.SetInt("isCameraMoved", 1);
+            Debug.Log("Game is focused");
+        }
+    }
+
+    private void OnApplicationPause(bool pauseStatus)
+    {
+        isGameOnPause = pauseStatus;
+
+        if (isGameOnPause)
+        {
+            PlayerPrefs.DeleteKey("isCameraMoved");
+            Debug.Log("Game is paused");
+        }
+        else
+        {
+            PlayerPrefs.SetInt("isCameraMoved", 1);
+            Debug.Log("Game is unpaused");
+        }
     }
 }
